pdf_to_word_V2/requirements.txt
pdf2docx>=0.5.6
PyMuPDF>=1.23.0
PyYAML>=6.0
python-docx>=0.8.11

